+++ termal_core/src/progress/bar_theme.rs
@@ -1,42 +1,60 @@
-pub type DefaultBarTheme = ();
+/// The default progress bar theme.
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash, Default)]
+pub struct DefaultBarTheme;
 
 pub trait BarTheme {
     /// The default theme for spaces.
-    fn reset(&self) -> &str;
-
-    /// The style of task name.
-    fn task(&self) -> &str;
-
-    /// The style of the percentage.
-    fn percent(&self) -> &str;
-
-    /// Style of the progress bar start.
-    fn start(&self) -> &str;
-
-    /// Style of the progress bar full.
-    fn full(&self) -> &str;
-
-    /// Style of the progress bar thumb.
-    fn thumb(&self) -> &str;
-
-    /// Style of the progress bar empty.
-    fn empty(&self) -> &str;
-
-    /// Style of the progress bar end.
-    fn end(&self) -> &str;
-
-    /// Style of the time.
-    fn time(&self) -> &str;
-
-    /// Style of the info.
-    fn info(&self) -> &str;
-}
-
-impl BarTheme for () {
     fn reset(&self) -> &str {
         "\x1b[0m"
     }
 
+    /// The style of task name.
+    fn task(&self) -> &str {
+        ""
+    }
+
+    /// The style of the percentage.
+    fn percent(&self) -> &str {
+        ""
+    }
+
+    /// Style of the progress bar start.
+    fn start(&self) -> &str {
+        "["
+    }
+
+    /// Style of the progress bar full.
+    fn full(&self) -> &str {
+        "#"
+    }
+
+    /// Style of the progress bar thumb.
+    fn thumb(&self) -> &str {
+        ""
+    }
+
+    /// Style of the progress bar empty.
+    fn empty(&self) -> &str {
+        " "
+    }
+
+    /// Style of the progress bar end.
+    fn end(&self) -> &str {
+        "]"
+    }
+
+    /// Style of the time.
+    fn time(&self) -> &str {
+        ""
+    }
+
+    /// Style of the info.
+    fn info(&self) -> &str {
+        ""
+    }
+}
+
+impl BarTheme for DefaultBarTheme {
     fn task(&self) -> &str {
         "\x1b[97m"
     }