+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## future
+## v5.0.0
 ### Breaking changes
 - The trait `IoProvider` now also requires that implementation of new trait
   `WaitForEvent`.